requirements.txt
# This project currently uses only the Python Standard Library.
# No external dependencies are required.
flet==0.80.3
fastapi
uvicorn
pydantic
google-genai